Multiple fixes buff size for GetText and SetText.
git-svn-id: http://svn.miranda-ng.org/main/trunk@11165 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/Quotes/src/ImportExport.cpp
@@ -171,7 +171,7 @@ namespace
 		LPTSTR p = pszBuffer;
 		LPCTSTR pszXml = TranslateT("XML File (*.xml)");
 		lstrcpyn(p,pszXml, (int)cBuffer);
-		size_t nLen = (int)lstrlen(pszXml)+1;
+		size_t nLen = lstrlen(pszXml)+1;
 		p+= nLen;
 		if(nLen < cBuffer)
 		{
